1.自我禁制修改参数类型

+++ app/src/main/java/org/cxct/sportlottery/ui/common/CustomPasswordVerifyDialog.kt
@@ -118,7 +118,7 @@ class CustomPasswordVerifyDialog : BaseDialog<SelfLimitViewModel>(SelfLimitViewM
             PassVerifyEnum.BET -> {
                 if (!checkPassword(password)) return
                 inputValue?.let {
-                    viewModel.passwordVerifyForLimitBet(MD5Util.MD5Encode(password), it.toInt())
+                    viewModel.passwordVerifyForLimitBet(MD5Util.MD5Encode(password), it.toDouble().toInt())
                 }
             }
         }
